refactor(gateway): extract adapters and schema into modules
Extract Telegram and LINE adapters from monolithic main.rs into
separate modules under gateway/src/adapters/. Extract shared
gateway event/reply schema types into gateway/src/schema.rs.

Pure refactor — zero behavior changes:
- Telegram adapter: webhook, reply, reactions, create_topic preserved
- LINE adapter: webhook, hybrid Reply/Push dispatch preserved
- All existing tests preserved and passing
- No OAB core changes
