+++ gateway/src/adapters/mod.rs
@@ -1,0 +1,2 @@
+pub mod line;
+pub mod telegram;
